5 - Quinto/Ingeniería Web/TP/requirements.txt
Django==1.11
ipython==5.3.0
pytz==2017.2
django-toolbelt
django-extensions
django-mptt
Pillow
git+https://github.com/django-haystack/django-haystack
whoosh
django-haystack-elasticsearch5
djangorestframework